Write a test to validate unique constraint of block and validator together
