+++ v1/confirmation_blocks/tests/confirmation_block.py
@@ -1,5 +1,5 @@
 from rest_framework.reverse import reverse
-from rest_framework.status import HTTP_200_OK, HTTP_201_CREATED
+from rest_framework.status import HTTP_200_OK, HTTP_201_CREATED, HTTP_400_BAD_REQUEST
 from thenewboston.third_party.pytest.asserts import assert_objects_vs_dicts
 
 
@@ -23,3 +23,18 @@ def test_confirmation_block_post(client, validator, block, confirmation_block_da
         expected=HTTP_201_CREATED,
     )
     assert confirmation_block_data['message']['block_identifier'] == response['block_identifier']
+
+
+def test_confirmation_block_post_400_unique_recipients(
+    client,
+    validator,
+    block,
+    confirmation_block_data_unique_recipients,
+):
+    response = client.post_json(
+        reverse('confirmationblock-list'),
+        confirmation_block_data_unique_recipients,
+        expected=HTTP_400_BAD_REQUEST,
+    )
+
+    assert response['message']['block']['non_field_errors'] == ['Tx recipients must be unique']
